Change pyldap->python-ldap in the README
Follow-up to 5455333d3252047c3e2436782830c6a9852a0f71 and
bd69525c3526e463b7071af2df254ed811999439.

The pyldap link pointed to an archived repository for a fork that has
been merged into mainline python-ldap package as of 3.0. This package
specifies python-ldap>=3.0.0 as dependency, so there's no need to
reference the out-of-date fork.